Colour code SCUBA-2 IDs by spec-z
I measured 250,350,500um fluxes on SPIRE maps for 21/27 SCUBA-2 sources in the HS1700 field. I estimated the uncertainties by the confusion + instrumental noises (from Kato+2016).

Here I show the estimated photo-z in comparison to their IRAC ID’s known redshift (photo-z denoted with a (p), otherwise spec-z). These of course are only comparable if they are the same source, so I put a "same source?" column (determined by the P-value). I say the sources statistically have the same redshift if the uncertainty in the photo-z is the same or greater than the difference between the photo-z and the known-z.
 
Summary:
- The results are somewhat in agreement (about 50% of them agree with their known redshift).
- There are cases where we know the source not to be at z=2.3 and we find the photo-z to be z=2.3+/-0.3 (e.g. 1700_7), however the uncertainty in the photo-z is way larger than the spec-z that told us z!=2.3, so this doesn’t mean much
- There are ~10 potential other sources that could be in the protocluster, though the uncertainty in the photo-z is far to large to say anything meaningful.
- I’m not too sure how useful this information is, I think the data is too scattered to tell us much, regardless if the photo-z’s agree with our known-z’s or not.